Changes: - Renamed python files - Formatted python files - Changed folder structure - Added a requirements.txt file - Added a .gitignore file - Removed __pycache__ folder
