+++ requirements.txt
@@ -1,0 +1,26 @@
+absl-py==1.4.0
+attrs==23.1.0
+black==23.7.0
+cffi==1.15.1
+click==8.1.6
+contourpy==1.1.0
+cycler==0.11.0
+flatbuffers==23.5.26
+fonttools==4.42.0
+kiwisolver==1.4.4
+matplotlib==3.7.2
+mediapipe==0.10.3
+mypy-extensions==1.0.0
+numpy==1.25.2
+opencv-contrib-python==4.8.0.74
+opencv-python==4.8.0.74
+packaging==23.1
+pathspec==0.11.2
+Pillow==10.0.0
+platformdirs==3.10.0
+protobuf==3.20.3
+pycparser==2.21
+pyparsing==3.0.9
+python-dateutil==2.8.2
+six==1.16.0
+sounddevice==0.4.6
